refactor(services): move ServiceKey/ParseServiceKey to registry
